refactor(rules-builder): split PredicateBuilder into cache, helpers, state
- predicateKeyCache.svelte.ts: module-scoped singleton for the
  compendium-tag-key scan (replaces the second <script module> block).
- predicateBuilderRows.ts: pure helpers (detectOperator, rowFromEntry,
  rowsFromValue, maybeNumber, rowToStatement, rowsToValue, emptyRow)
  plus the Operator/RowState/OPERATORS types and constant.
- PredicateBuilderState.svelte.ts: per-instance reactive state factory
  (rows, suggestionKeys, preview, datalistId + handlers and the
  value-sync effect).

The .svelte file now has a single <script> tag holding only imports,
prop destructure, state creation, and an onMount call to warm the
cache.
